% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{conv_with_kernel}
\alias{conv_with_kernel}
\title{Apply 2D convolution on a numeric vector}
\usage{
conv_with_kernel(vals, kernel_vec)
}
\arguments{
\item{vals}{Numeric vector (from focal neighborhood)}

\item{kernel_vec}{Numeric vector (flattened kernel)}
}
\value{
Numeric single value
}
\description{
Internal helper for Sobel kernels.
}
\keyword{internal}
