% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_fill_raster.R
\name{fill_ground_raster}
\alias{fill_ground_raster}
\title{Fill raster values using neighborhood-based ground estimation}
\usage{
fill_ground_raster(x, w = 15)
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{w}{Focal window size (odd integer).}
}
\value{
A \code{terra::SpatRaster}.
}
\description{
Applies a focal operation to fill NA values in a raster using
a clustering-based ground estimation method.
}
